+++ app/Questions.py
@@ -4,7 +4,6 @@ import itertools
 from abc import ABC, abstractmethod, abstractproperty
 import json
 import uuid
-from StorageHandler import write_to_file, load_file_as_json
 import copy
 
 
@@ -224,49 +223,3 @@ class FillInTheBlankQuestion(Question):
 # This should be decoupled from the storage(or not needed by the storage). Quiz id's should be unique as well. This
 # logic will need to thought out a little more. I can see issues with deleting quizzes but still needing to keep them
 # alive when there's a reference to it in client Code. Essentially, this could become an issue on the client side?
-
-class Quiz:
-
-    # I did not initialize with an empty list because doing so will result in the same list object being passed to all
-    # new objects.
-
-    @staticmethod
-    # TODO find a way to ensure multiple of the same quiz are not loaded.
-    def load_quiz_from_json(json_object):
-        return Quiz(json_object['name'], questions=[Question.create_a_question_from_json(question) for question in json_object['questions']], id=json_object['id'])
-
-    def __init__(self, name: str, questions: List[Question] = None, id=None):
-        self.name = name
-        if questions is None:
-            self.questions = []
-        else:
-            self.questions = questions
-        if id is None:
-            self.id = str(uuid.uuid4())
-        else:
-            self.id = id
-
-    def add_question_to_quiz(self, question: Question):
-        if not isinstance(question, Question):
-            raise ValueError("Only questions can be added to the instance variable questions!")
-        self.questions.append(question)
-
-    def get_question(self, question_id):
-        return self.questions[question_id]
-
-    def jsonify(self):
-        return json.dumps(
-            {'id': self.id, 'name': self.name, 'questions': [json.loads(question.jsonify()) for question in self.questions]})
-
-
-if __name__ == '__main__':
-    quiz = Quiz("Brian's First Quiz")
-    multiple_choice_question = MultipleChoiceQuestion(prompt="Who is the best?", choices={"A": 'Mike', "B": 'Domingo'},
-                                                      answer='A')
-    quiz.add_question_to_quiz(multiple_choice_question)
-    write_to_file(quiz.jsonify(), 'test.json')
-    quiz = Quiz.load_quiz_from_json((load_file_as_json('test.json')))
-    first_question = quiz.get_question(0)
-    first_question.add_response('A')
-    first_question.add_response('B')
-    write_to_file(quiz.jsonify(), 'taken_quiz.json')
